Sun, Jun 26, 2022, 1:14 PM +00:00

+++ pset1/credit/credit1.c
@@ -75,7 +75,7 @@ bool checksum(int sum3)
     }
 }
 
-int numlen(int card)
+int numlen(unsigned long long card)
 {
     int length = 1;
     while (card >= 10)
